% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{dashtemplate}
\alias{dashtemplate}
\alias{dashtemplate-package}
\alias{nameyourfunction}
\title{Dashboard Template}
\usage{
nameyourfunction(data, params, ...)
}
\arguments{
\item{data}{A list of datasets retrieved from PiLR}

\item{params}{A list of all arguments sent to pilr_dpu}
}
\description{
This is a template for creating functions to be used as PiLR DPUs.
}

